requirements.txt
# Production Dependencies
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
google-genai>=0.8.0
pypdf2>=3.0.1
python-docx>=1.1.0
python-multipart>=0.0.9
psutil>=5.9.0
